Add permission middleware

+++ membership/lumen/routes/web.php
@@ -14,22 +14,23 @@
 $app->  post("membership/authenticate", "Member@authenticate");   // Authenticate a member
 
 // Members
-$app->   get("membership/member",      "Member@list");   // Get collection
-$app->  post("membership/member",      "Member@create"); // Model: Create
-$app->   get("membership/member/{id}", "Member@read");   // Model: Read
-$app->   put("membership/member/{id}", "Member@update"); // Model: Update
-$app->delete("membership/member/{id}", "Member@delete"); // Model: Delete
-$app->   get("membership/member/{id}/groups", "Member@getGroups");    // Get collection with members
-$app->  post("membership/member/{id}/groups/add", "Member@addGroup");    // Get collection with members
-$app->  post("membership/member/{id}/groups/remove", "Member@removeGroup");    // Get collection with members
+$app->   get("membership/member",      ['middleware' => 'permission:member_view',   'uses' => "Member@list"]);   // Get collection
+$app->  post("membership/member",      ['middleware' => 'permission:member_create', 'uses' => "Member@create"]); // Model: Create
+$app->   get("membership/member/{id}", ['middleware' => 'permission:member_view',   'uses' => "Member@read"]);   // Model: Read
+$app->   put("membership/member/{id}", ['middleware' => 'permission:member_edit',   'uses' => "Member@update"]); // Model: Update
+$app->delete("membership/member/{id}", ['middleware' => 'permission:member_delete', 'uses' => "Member@delete"]); // Model: Delete
+$app->   get("membership/member/{id}/permissions",   ['middleware' => 'permission:permission_manage',   'uses' => "Member@getPermissions"]);    // Get a member's permissions
+$app->   get("membership/member/{id}/groups",        ['middleware' => 'permission:group_member_view',   'uses' => "Member@getGroups"]);    // Get collection with members
+$app->  post("membership/member/{id}/groups/add",    ['middleware' => 'permission:group_member_add',    'uses' => "Member@addGroup"]);    // Get collection with members
+$app->  post("membership/member/{id}/groups/remove", ['middleware' => 'permission:group_member_remove', 'uses' => "Member@removeGroup"]);    // Get collection with members
 
 // Groups
-$app->   get("membership/group",       "Group@list");    // Get collection
-$app->  post("membership/group",       "Group@create");  // Model: Create
-$app->   get("membership/group/{id}",  "Group@read");    // Model: Read
-$app->   put("membership/group/{id}",  "Group@update");  // Model: Update
-$app->delete("membership/group/{id}",  "Group@delete");  // Model: Delete
-$app->   get("membership/group/{id}/members", "Group@getMembers");    // Get collection with members
+$app->   get("membership/group",       ['middleware' => 'permission:group_view',   'uses' => "Group@list"]);    // Get collection
+$app->  post("membership/group",       ['middleware' => 'permission:group_create', 'uses' => "Group@create"]);  // Model: Create
+$app->   get("membership/group/{id}",  ['middleware' => 'permission:group_view',   'uses' => "Group@read"]);    // Model: Read
+$app->   put("membership/group/{id}",  ['middleware' => 'permission:group_edit',   'uses' => "Group@update"]);  // Model: Update
+$app->delete("membership/group/{id}",  ['middleware' => 'permission:group_delete', 'uses' => "Group@delete"]);  // Model: Delete
+$app->   get("membership/group/{id}/members", ['middleware' => 'permission:group_member_view',  'uses' => "Group@getMembers"]);    // Get collection with members
 
 // Roles
 $app->   get("membership/role",        "Role@list");     // Get collection
@@ -39,8 +40,10 @@ $app->   put("membership/role/{id}",   "Role@update");   // Model: Update
 $app->delete("membership/role/{id}",   "Role@delete");   // Model: Delete
 
 // Permissions
-$app->   get("membership/permission",        "Permission@list");     // Get collection
-$app->  post("membership/permission",        "Permission@create");   // Model: Create
-$app->   get("membership/permission/{id}",   "Permission@read");     // Model: Read
-$app->   put("membership/permission/{id}",   "Permission@update");   // Model: Update
-$app->delete("membership/permission/{id}",   "Permission@delete");   // Model: Delete
+$app->   get("membership/permission",      ['middleware' => 'permission:permission_manage',  'uses' => "Permission@list"]);     // Get collection
+$app->  post("membership/permission",      ['middleware' => 'permission:permission_manage',  'uses' => "Permission@create"]);   // Model: Create
+$app->   get("membership/permission/{id}", ['middleware' => 'permission:permission_manage',  'uses' => "Permission@read"]);     // Model: Read
+$app->   put("membership/permission/{id}", ['middleware' => 'permission:permission_manage',  'uses' => "Permission@update"]);   // Model: Update
+$app->delete("membership/permission/{id}", ['middleware' => 'permission:permission_manage',  'uses' => "Permission@delete"]);   // Model: Delete
+
+$app->  post("membership/permission/register", "Permission@batchRegister");
